+++ structs/__init__.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 
-version_info = (0, 0, 1)
+version_info = (0, 0, 2)
 
 __title__ = 'structs'
 __author__ = 'Jon Nappi'
 __version__ = '.'.join([str(x) for x in version_info])
-__build__ = 0x000001
+__build__ = 0x000002
 __license__ = 'Apache 2.0'
 __copyright__ = 'Copyright 2015 Jonathan Nappi'