+++ HISTORY.rst
@@ -1,5 +1,12 @@
 Release History
 ---------------
+0.0.2 (2015-01-24)
+++++++++++++++++++
+
+* Added arrays.ParallelArray
+* Added arrays.OrganizedList
+* Improved Unit level testing (up to 99% coverage)
+
 0.0.1 (2015-01-19)
 ++++++++++++++++++